requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_0
chardet=3.0.4=pypi_0
click=8.0.3=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cython=0.29.24=pypi_0
detectron2=0.6+cu101=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20211023=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.42.0=pypi_0
hydra-core=1.1.1=pypi_0
idna=2.10=pypi_0
importlib-metadata=4.8.2=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=py38hd09550d_3
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.1.1=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
pathspec=0.9.0=pypi_0
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
portalocker=2.3.2=pypi_0
protobuf=3.19.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pydot=1.4.2=pypi_0
python=3.8.12=h12debd9_0
pytorch=1.8.0=py3.8_cuda10.2_cudnn7.6.5_0
readline=8.1=h27cfd23_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu102
tqdm=4.62.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.25.11=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
